Ui Fixes, Logger fix, PostsDAO singel item bug fix
Implemented columns in the post item and the user manager nav bar. The
‘contact user’ lightwindow UI was fixed also. The welcome box text was
updated. I also uncommitted all the log commands in the models and lib
and everywhere because the log bug was fixed. Also, the bug where the
PostsDAO returns a single object when a single item is queried was
fixed, instead the postsDAO returns an array with just one item
instead, and the application was modified though-out to handle this.